Made changes to genjava file for external classes file and output path
